05.springsecurity/02.security-filter/src/main/java/me/kickscar/spring/security/config/configurer/SecurityConfig02.java
package me.kickscar.spring.security.config.configurer;

import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.security.config.annotation.web.builders.HttpSecurity;
import org.springframework.security.config.annotation.web.builders.WebSecurity;
import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
import org.springframework.security.web.SecurityFilterChain;
import org.springframework.security.web.util.matcher.AntPathRequestMatcher;

@Configuration
@EnableWebSecurity
public class SecurityConfig02 {
    public WebSecurityCustomizer webSecurityCustomizer() {
        return new WebSecurityCustomizer() {
            @Override
            public void customize(WebSecurity web) {
                web
                        .ignoring()
                        .requestMatchers(new AntPathRequestMatcher("/assets/**"));
            }
        };
    }

    @Bean
    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {

        /*
             HttpSecurity:
             is similar to Spring Security's XML <http> element in the namespace configuration.

             It allows configuring web based security for specific http requests.
             By default it will be applied to all requests, but can be restricted using #requestMatcher(RequestMatcher) or other similar methods.
         */

        return http.build();
    }
}
